refactor(safety-authority): ADR-0035 Stage 3a — extract the pure posture-decision surface
The leading-edge slice of Stage 3 (decomposing the `AppState` god-object into a
`kirra-safety-authority` crate). Per the Stage-3 risk register the DAG traversal
and actuator core move LAST, behind the full Kani/loom/power-loss/replay net; this
first slice moves ONLY the pure, AppState-free safety-DECISION surface to prove the
crate / shim / MSRV / guardrail mechanics:

- new `crates/kirra-safety-authority` (deps: kirra-core + serde) holding
  `FleetNodePosture`, `derive_fleet_posture` (the fail-closed fleet-fold),
  `RssRecoveryStreak`, and `LockoutReason` (+ its Display code mapping).
- each original location keeps a `pub use kirra_safety_authority::…` shim, so every
  `crate::verifier::FleetNodePosture` / `crate::posture_engine::derive_fleet_posture`
  / `crate::posture_engine_v2::LockoutReason` path resolves unchanged. No consumer
  is edited; INVARIANT #11 (`State<Arc<ServiceState>>`, `svc.app.*`) untouched.
- serde `rc` feature enabled on the new crate (FleetNodePosture interns ids as
  `Arc<str>`; serde only impls Serialize/Deserialize for Arc/Rc under `rc`).

Behaviour-preserving: the moved items are byte-identical; the root's own
derive_fleet_posture / DAG / lockout / verifier tests exercise them through the
shims (root lib 599/0 locally). No behaviour change — pure relocation.
